updated creation of invitation, application, platform community to set parent directly on child; split community role module out from community module
