+++ src/domain/community/community-role/community.role.lifecycle.application.options.provider.ts
@@ -9,17 +9,17 @@ import { MachineOptions } from 'xstate';
 import { LifecycleService } from '@domain/common/lifecycle/lifecycle.service';
 import { ApplicationService } from '@domain/community/application/application.service';
 import { EntityNotInitializedException } from '@common/exceptions';
-import { CommunityService } from './community.service';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { AuthorizationService } from '@core/authorization/authorization.service';
 import { AuthorizationPolicy } from '@domain/common/authorization-policy';
 import { CommunityRole } from '@common/enums/community.role';
+import { CommunityRoleService } from './community.role.service';
 
 @Injectable()
-export class CommunityApplicationLifecycleOptionsProvider {
+export class CommunityRoleApplicationLifecycleOptionsProvider {
   constructor(
     private lifecycleService: LifecycleService,
-    private communityService: CommunityService,
+    private communityRoleService: CommunityRoleService,
     private authorizationService: AuthorizationService,
     private applicationService: ApplicationService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER) private readonly logger: LoggerService
@@ -77,7 +77,7 @@ export class CommunityApplicationLifecycleOptionsProvider {
             LogContext.COMMUNITY
           );
 
-        await this.communityService.assignUserToRole(
+        await this.communityRoleService.assignUserToRole(
           community,
           userID,
           CommunityRole.MEMBER,
