+++ src/domain/community/community-role/dto/community.role.dto.apply.ts
@@ -6,7 +6,7 @@ import { Type } from 'class-transformer';
 import { UUID_LENGTH } from '@common/constants';
 
 @InputType()
-export class CommunityApplyInput {
+export class CommunityRoleApplyInput {
   @Field(() => UUID, { nullable: false })
   @MaxLength(UUID_LENGTH)
   communityID!: string;
