+++ src/domain/community/community-role/index.ts
@@ -1,0 +1,2 @@
+export * from './dto/community.role.dto.role.assign.user';
+export * from './dto/community.role.dto.role.remove.user';
