+++ src/domain/community/community/index.ts
@@ -1,4 +1,2 @@
 export * from './community.entity';
 export * from './community.interface';
-export * from './dto/community.dto.role.assign.user';
-export * from './dto/community.dto.role.remove.user';
