+++ src/domain/space/space/space.service.ts
@@ -60,6 +60,7 @@ import { SpaceLevel } from '@common/enums/space.level';
 import { UpdateSpaceSettingsInput } from './dto/space.dto.update.settings';
 import { IContributor } from '@domain/community/contributor/contributor.interface';
 import { CommunityContributorType } from '@common/enums/community.contributor.type';
+import { CommunityRoleService } from '@domain/community/community-role/community.role.service';
 
 @Injectable()
 export class SpaceService {
@@ -69,6 +70,7 @@ export class SpaceService {
     private contextService: ContextService,
     private agentService: AgentService,
     private communityService: CommunityService,
+    private communityRoleService: CommunityRoleService,
     private namingService: NamingService,
     private profileService: ProfileService,
     private spaceSettingsService: SpaceSettingsService,
@@ -848,7 +850,7 @@ export class SpaceService {
     // Before assigning roles in the subspace check that the user is a member
     if (agentInfo) {
       const agent = await this.agentService.getAgentOrFail(agentInfo?.agentID);
-      const isMember = await this.communityService.isMember(
+      const isMember = await this.communityRoleService.isMember(
         agent,
         space.community
       );
@@ -905,7 +907,7 @@ export class SpaceService {
       );
     }
 
-    await this.communityService.assignContributorAgentToRole(
+    await this.communityRoleService.assignContributorAgentToRole(
       space.community,
       contributor.agent,
       role,
@@ -924,21 +926,21 @@ export class SpaceService {
       );
     }
     if (agentInfo) {
-      await this.communityService.assignUserToRole(
+      await this.communityRoleService.assignUserToRole(
         space.community,
         agentInfo.userID,
         CommunityRole.MEMBER,
         agentInfo
       );
 
-      await this.communityService.assignUserToRole(
+      await this.communityRoleService.assignUserToRole(
         space.community,
         agentInfo.userID,
         CommunityRole.LEAD,
         agentInfo
       );
 
-      await this.communityService.assignUserToRole(
+      await this.communityRoleService.assignUserToRole(
         space.community,
         agentInfo.userID,
         CommunityRole.ADMIN,
@@ -1007,6 +1009,7 @@ export class SpaceService {
 
     await this.contextService.removeContext(space.context.id);
     await this.collaborationService.deleteCollaboration(space.collaboration.id);
+    await this.communityRoleService.removeAllCommunityRoles(space.community);
     await this.communityService.removeCommunity(space.community.id);
     await this.profileService.deleteProfile(space.profile.id);
     await this.agentService.deleteAgent(space.agent.id);
@@ -1259,7 +1262,7 @@ export class SpaceService {
 
   public async getMembersCount(space: ISpace): Promise<number> {
     const community = await this.getCommunity(space.id);
-    return await this.communityService.getMembersCount(community);
+    return await this.communityRoleService.getMembersCount(community);
   }
 
   public async getPostsCount(space: ISpace): Promise<number> {
@@ -1309,7 +1312,9 @@ export class SpaceService {
     const community = await this.getCommunity(space.id);
 
     // Members
-    const membersCount = await this.communityService.getMembersCount(community);
+    const membersCount = await this.communityRoleService.getMembersCount(
+      community
+    );
     const membersTopic = new NVP('members', membersCount.toString());
     membersTopic.id = `members-${space.id}`;
     metrics.push(membersTopic);
