+++ src/services/api/conversion/conversion.module.ts
@@ -3,11 +3,12 @@ import { ConversionService } from './conversion.service';
 import { ConversionResolverMutations } from './conversion.resolver.mutations';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { SpaceModule } from '@domain/space/space/space.module';
-import { CommunityModule } from '@domain/community/community/community.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
 import { CommunicationModule } from '@domain/communication/communication/communication.module';
 import { AccountModule } from '@domain/space/account/account.module';
 import { NamingModule } from '@services/infrastructure/naming/naming.module';
+import { CommunityRoleModule } from '@domain/community/community-role/community.role.module';
+import { CommunityModule } from '@domain/community/community/community.module';
 
 @Module({
   imports: [
@@ -15,6 +16,7 @@ import { NamingModule } from '@services/infrastructure/naming/naming.module';
     AccountModule,
     SpaceModule,
     CommunityModule,
+    CommunityRoleModule,
     AuthorizationModule,
     AuthorizationPolicyModule,
     NamingModule,
