+++ src/services/api/me/me.module.ts
@@ -10,8 +10,8 @@ import { SpaceModule } from '@domain/space/space/space.module';
 import { RolesModule } from '../roles/roles.module';
 import { ActivityLogModule } from '../activity-log/activity.log.module';
 import { ActivityModule } from '@platform/activity/activity.module';
-import { CommunityModule } from '@domain/community/community/community.module';
 import { EntityResolverModule } from '@services/infrastructure/entity-resolver/entity.resolver.module';
+import { CommunityRoleModule } from '@domain/community/community-role/community.role.module';
 
 @Module({
   imports: [
@@ -23,7 +23,7 @@ import { EntityResolverModule } from '@services/infrastructure/entity-resolver/e
     RolesModule,
     ActivityLogModule,
     ActivityModule,
-    CommunityModule,
+    CommunityRoleModule,
     EntityResolverModule,
   ],
   providers: [MeService, MeResolverQueries, MeResolverFields],
